Use feature unix_sigpipe to avoid panic on pipe close
Set `SIGPIPE` to `SIG_DFL`.

Without this, when the pipe is closed before the rust binary is finished
writing, or if the rust binary writes more than `PIPE_MAX_SIZE` bytes
before the receiver has read enough bytes to make space in the pipe,
process panics with:

    # stable-x86_64-unknown-linux-gnu unchanged -
    # rustc 1.72.0 (5680fa18f 2023-08-23)
    thread 'main' panicked at 'failed printing to stdout: Broken pipe
    (os error 32)', library/std/src/io/stdio.rs:1019:9

    # nightly-x86_64-unknown-linux-gnu unchanged -
    # rustc 1.75.0-nightly (aa1a71e9e 2023-10-26)
    thread 'main' panicked at library/std/src/io/stdio.rs:1021:9:
    failed printing to stdout: Broken pipe (os error 32)

With this, the process exits normally with SIGPIPE (13).

References:
  - https://github.com/rust-lang/rust/issues/97889
  - https://github.com/rust-lang/rust/issues/62569
  - https://unix.stackexchange.com/a/626976/7688
  - https://unix.stackexchange.com/a/433348/7688